requirements.txt
asgiref==3.4.1
black==21.10b0
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
colorama==0.4.4
dj-database-url==0.5.0
Django==3.2.9
django-heroku==0.3.1
djangorestframework==3.12.4
gunicorn==20.1.0
idna==3.3
mypy-extensions==0.4.3
oauthlib==3.1.1
pathspec==0.9.0
platformdirs==2.4.0
psycopg2==2.9.1
pytz==2021.3
regex==2021.11.2
requests==2.26.0
requests-oauthlib==1.3.0
sqlparse==0.4.2
tomli==1.2.2
tweepy==4.3.0
typing-extensions==3.10.0.2
urllib3==1.26.7
whitenoise==5.3.0
